Overhauled the inheritance structure to reduce redundant code.
We introduce the Vector and Annotated classes as virtual bases for all
vector-like and annotated objects (duh) respectively. This is taken directly
from Bioconductor's inheritance hierarchy, and allows us to discard a lot of
the redundant methods for IRanges and GRanges. Similarly, DataFrame is rebased
onto Annotated (though not Vector, as the 'vector-ish' dimension for DFs is
context-dependent, so we don't want to bake that into the class definition).

We also rework the _bioconductor_* methods so as to allow a gradual build-up of
each new class instance via the parent methods. This avoids having to duplicate
the parent operations in each child class. We removed some of the modify-in-place
options because they would complicate the implementation.