Added a SummarizedExperiment class for handling assay data. (#4)
This involves introducing another suite of generics for operating on
two-dimensional objects, e.g., NUMBER_OF_ROWS, SLICE_2D, COMBINE_COLUMNS. The
idea is to avoid complicated overloads with the 1D suite of generics; it may
also allow objects to behave differently in 1D vs 2D contexts.

We also added a dense matrix class so that we have _something_ to put in the 
assays of our SE (for testing and examples if nothing else). It is likely that
this will be replaced by more efficient representations in actual applications,
but at least we have a simple baseline for testing SE and other classes.