Implemented our own version of a GRangesList. (#5)
This implementation is called GroupedGRanges, which is a nicer terminology as
we can talk about groups of genomic ranges as being elements of the
'GroupedGRanges vector'. Like BioC, we still use a single underlying GRanges,
which avoids a large overhead per group when we have lots of groups; this comes
at the cost of a slightly more complicated implementation.