Allow users to set individual groups in a GroupedGRanges.
This avoids a full reindexing of the internal concatenated representation by
staging $setGroup instructions and only executing them when getters are called
on the (ostensibly) modified object. In this manner, users can go wild with
calling $setGroup as many times as they want - as long as they do so
consecutively - and only pay the cost of reindexing once they're done.