Switch from Object.create(<class>.prototype) to new <class> in generics.
This avoids some previously observed problems with instances of the
former not being bound to private methods and fields. As such, we can
reintroduce the use of private methods/fields in various classes,
namely the GroupedGRanges and DenseMatrix.

However, it also means that we need to add no-op clauses to each
constructor so that, when it is invoked with no arguments, it returns an
empty object for filling by each generic. This further requires
modification of the Annotated constructor so that the metadata is
required (allowing us to identify the no-op invocation).