Enable combining of RSEs with GRanges, GroupedGRanges and plain SEs. (#10)
When multiple objects are in the mix to be combined by RSE's COMBINE_ROWS
method, the presence of a GroupedGRanges in the rowRanges of any object will
force any GRanges to be promoted to GroupedGRanges with length-1 groups.  Any
base SE will also be promoted to an RSE containing a GroupedGRanges with
length-0 groups (which in turn will trigger additional GRanges->GroupedGRanges
promotions). This ensures that the output is either a GRanges, if all objects
are RSEs containing GRanges; and a GroupedGRanges otherwise.