Split out the name lookup logic into a separate class.
This allows us to be more rigorous in deciding whether a new lookup table needs
to be created based on the modification to the names. It also gives us the
opportunity for some perf improvements by avoiding unnecessary lookup resets
for in-place operations that don't affect existing entries in the lookup.