Refactored generics to avoid passing around an empty object.
Now, each _bioconductor_* method just returns an instance of the same class,
which is simpler than accepting an empty object and modifying it in-place.

In practice, we still create empty objects inside the methods of the base
class, so that methods for subclasses can fill in subclass-specific bits; but
that's an implementation detail rather than part of the method signature.

Also mention that base classes expect their subclasses to have constructors
that work with no arguments, so that an empty object can be properly created.