Store Annotated metadata as a List instead of a Map.
This allows us to handle unnamed metadata and duplicate keys.

We also fix the List constructor so that a valid object is created when called
with no arguments. This allows users to call 'new List' and add elements later,
which is a common pattern for list-like containers. (Other classes don't need
to have sane empty objects because those empty objects are kind of useless.)